ml_service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pymupdf==1.25.4
scikit-learn==1.6.1
sentence-transformers==4.0.1
joblib==1.4.2
google-generativeai==0.8.4
python-multipart==0.0.20
python-dotenv==1.1.0
numpy==2.2.4
